[arch/x86/time/rtc.c] The RTC should no longer be used a system timer. It's default frequency is now 2HZ and it's periodic interrupt is only used to set the date.
Separate the date reading code into it's own function, as it's now being
used to read the date initially instead of waiting for the first
interrupt to occur.

The RTC needs further work done on it. Like making it able to generate
timer tick interrupts in case someone wants that.