[kconfig] Tweak kernel configurations.
Change the name of some configuration options:
  - All config options are now prefixed with DXGMX_CONFIG_
  - Remove '_' from the begining and end of DXGMX_(VER_MAJ/VER_MIN/PATCH_N/CODENAME) macros.

Add a separate makefile-like file in scripts/ named configparser.mk that takes care of confiration parsing instead of just doing it in the root Makefile.

Officially support the use of both numbers and names when specifying DXGMX_CONFIG_LOG_LEVEL and add the QUIET log level to the KLogLevel enum.

Update documentations to reflect the above changes.