[module] Add basic implementation for kernel modules.
We can now compile built-in modules into the kernel to provide optional
features.

The way this works is: there is a new section in the kernel binary
called '.modules'. This section gets populated with Module structs that
have the MODULE specifier. They are then loaded by the
module_load_builtins() function.

This feature needs a lot of work, like:
  - Right now modules are built in the same way as the rest of
    the kernel. I think it would be nice to have a way of saying that a
    module is actually a module instead of some arbitrary piece of
    kernel code.
  - When we *get* to user-space it would be nice to let the user load
    modules.
  - We can definitely add more information/functionality about/to each module.