[build-system] Add example TOOLCHAIN and TARGET files in toolchain/.
The TOOLCHAIN file specifies an i686 LLVM toolchain (with GNU objcopy because
llvm-objcopy doesn't suppport --update-section yet) and links against
clang_rt.builtins which is specific to my machine and should be
replaced as needed.

The TARGET file builds a 'debug' version with maximum verbosity, no
optimizations, debug symbols and a few modules.