[vfs.c] Add vfs_open(), vfs_read() and vfs_close().
vfs_open() opens a file and returns a file descriptor number.

vfs_read() uses the file descriptor number of an opened file
to read the contents of said file.

vfs_close() should close the file, but is a stub for now.

All of the above functions are very primitive, and while they do kind of
work, they have a long way to go before they are remotely useful.