[vnode] Replace the 'parent' field with the parent's 'parent_n'.
This is because I realized that if we reallocate a vnode vector (as we
usually do when adding or removing files), the vector's base address might
change, leaving any vnodes who have a parent directory with a pointer to
a now invaid vnode.

This way of doing it means we have to walk the vnode table when we want
to find a parent, which is obviously pretty slow. This could be fixed by
using a hashtable or a linked list of sorts, but I can't be fucked right
now. :)

As side note: We may want to take the responsability of creating vnodes
off of the filesystem driver and instead have the vfs? create vnodes for
the given filesystem.