[syscalls-x86] Add basic implementation of system calls.
Altough not very useful at the moment, we can now issue system calls
from ring 3. The ABI for syscalls is inspired by Linux, where the
interrupt's vector number is 0x80, the syscall number is stored in eax,
and the arguments follow in the ebx, ecx, edx, esi and edi registers.
Unlike Linux the ebp register is not used for a 6th argument because
we are still dependant on it for stack frames in order to generate call
stack backtraces. After an 'int $0x80' has been issued, when the control
is given back to the application the system call's return value will be
placed in the eax register. Note that if the system call has no return
value (void) the value 0 will be placed in eax.

There are two system call stubs at the moment for exit (1) and read (2),
if called they just print a message and panic the kernel.