[falloc] Change the way pageframes are allocated & tweak API.
We added a new way of requesting specific page frames, using their base.
This is useful if we are trying to map the kernel heap to physical
memory which should be done in a higher-half fashion, with pageframes
starting right after the kernel image.

The other way of allocating a pageframe is using falloc_one_user() which
returns the address of an available pageframe suitable for basically
anything else except the kernel heap (where pages should be contigous in
physical memory and easy to translate into physical addresses). Also
these pageframes are allocated starting from "high-addresses" growing
downwards, this is to "leave" space for the "low-address", which are
more likely to be used for the kernel heap. This is because the kernel
is loaded at a low address in memory with (optimally) no available
pageframes coming before it. The direction in which these two "types"
of pageframes could grow might look something like this:

            These are allocated manually
                         |
[ kernel_image ] [ lo pageframes ] --->      <--- [ hi pageframes ]
                                                          |
                                These are returned by falloc_one_user

This commit does not compile!