[kmalloc] Rewrite kmalloc.
Transform kmalloc into more of a front-end for whatever driver we
are using. Right now, the only driver available is "gallocator", which is
hardcoded in kmalloc_init(). I think it would be a good ideea to get rid
of this hardcode and implement multi-level modules (something similar to
linux maybe), that get initialized at different key-points in the kernel
boot process.

This commit does not compile!