kernel: Add kinit_stage3.
kinit_stage3 is a single function that acts like a userspace main(),
baked into the kernel image in the .kinit_stage3_text section.

The kernel spawns this binary as PID 1, and then the binary execs the
actual PID 1 using a syscall. We do this because the kernel doesn't have
the API to directly read from disks, without syscalls.

Code for loading into memory is to come.

This is inspired by how linux does things.