errno: Remove the errno variable.
The errno variable was only being used in a few places around the
kernel, and most of them have been easily replaced by just return
values. Plus I find working with errno to be a pain.

The only place where the errno variable made things easier was in
strtoul and strtol which rely on the errno variable to signal an error.
We replaced that behaviour by returning the errno value and taking in a
pointer to the destination value.