vfs: Cleanup + various fixes.
There are a lot of changes in this commit, that really should have been
separated into multiple ones, but the vfs was such a mess that I just
gave up.

First, A FileSystem now holds a LinkedList of VirtualNodes. This fixes a
problem with the previous way of doing it (arrays), in which, because these
VirtualNodes are referenced in the code using pointers,
calling krealloc on that array and possibly moving it's starting
address, would leave such references dangling.

Second, We now register FileSystemDrivers by taking in a pointer to
them, and storing that pointer. Same story as above.

Third, Move one function from "fs" into it's own API called "path".

Fourth, Separate some function pointers for working on VirtualNodes from
FileSystemDriver, into it's own struct called VirtualNodeOperations.

Fifth, General cleanup around the vfs, fs and vnode.

This leaves us with the 2 fs drivers we have: fat and ramfs, which are
still horrible, especially fat.