blkdevmanager -> blkdevm: Rename + cleanup.
Rename the blkdevmanager API to just blkdevm.

Also cleanup said API, that being said, I still feel like it's pretty
awkward. I think that's because I don't really a clue on how it's
supposed to be structured, since I don't know how it's gonna be
accessed. We barely have a userspace, and our ATA driver is
*poor* at best, so we're just gonna add to it as we go.