procm: Fix kernel stack bug.
In procm_replace_proc, we would kfree the acting's process
kernel stack, that being the kernel stack we were using. We now keep the
kernel stack when replacing a process.

Also fix a subtle bug in which we were freeing the acting process'
paging structure just before setting it to the new one. Problem was that
if an IRQ came between these two operations, and that IRQ called kmalloc
there would've been a chance, the just free'd paging structure would get
fucked.