types: Add ERR_OR macro.
This is a new experimental feature, allowing functions to return a value
along with an error code in a single struct, removing the need for just
returning an error code, and taking in a pointer to the argument to be
set.

The syntax looks like this:

/* Only needs to be defined once, and included wherever it's used. */
DEFINE_ERR_OR(float);

ERR_OR(float) func(float a, float b)
{
	if(b == 0)
		return ERR(float, -1);

	return VALUE(float, a / b);
}

Now any call to this function will look something like:

ERR_OR(float) res = func(1, 2);
if(res.error)
	exit(res.error);
else
	res.value;

As you can see, this thing is really just a stab at creating
static polymorphism in a language that doesn't support it. I don't
intend to take this any further, but I really hate the way you can't
really return an error and a value in a *non awkward* way. And this is
my best shot at fixing that.

There are of course a few weird things about it:
1: Any data types that ERR_OR will use, will first need to be defined
   with DEFINE_ERR_OR().

2: When returning a value or an error you have to specify the type of
   the value. ex: VALUE(float, 0) instead of VALUE(0).

3: Right now we can't to stuff like ERR_OR(float*), as internally the
   macro declares a struct with the stringified type name, and of
   course struct S_float* {...} is not a valid declaration. There are
   ways of fixing this, but it involves creating more macros and more
   awkwardness.

Again, this is experimental, and might very well get removed later if it
proves to do more bad than good.