procm: Fix procm_replace_proc leaving us without a valid paging struct.
When trying to free the context of the process we were trying to
replace, we would effectively free the paging struct that the system was
using at that time. We now call procm_load_ctx() with the new process
before freeing the old context so we are safe from stuff like that.

This was a fun one to debug :)