procm: Add procm_spawn_kernel_proc().
This dummy temporary process is spawned because we will need to access
the vfs for mounting/reading from the future initrd. This process is
replaced by PID 1 when the time comes. This is basically a
chicken-and-egg problem where we need to access the vfs API to execute
PID 1 but the vfs API can only be accessed from userspace apps, because
it works on a process context. This patch is a hack but I think it's the
best hack given our problem.