user: Add user_copy_to().
user_copy_to() is used to copy data from kernel space to userspace in a
safe and predictable way. The way it works is inspired by linux's
implementation of this function's equivalent, after trying a bunch of
alternatives which either failed miserably, or just seemed *weird*.

user_copy_to sits in it's own elf section. When invoked it starts
copying data. If a fault occurs while copying the fault handler looks at
the faulting eip and if it falls inside the user_copy_to region it
returns control to a "function" that cleans up the stack and returns
-EFAULT to the caller.

NOTES:

Right now the function doesn't really care if the fault is because of
the kernel src buffer or the userspace dest buffer. (Should we care?)

Upon a fault the page fault handler immediately jumps to the clean-up
function. In the future we might want to do some stuff before resorting
to that.

This whole thing was implemented after reading this stackoverflow
answear:
https://stackoverflow.com/questions/8265657/how-does-copy-from-user-from-the-linux-kernel-work-internally