syscalls: Refactor syscalls.
Created an API for defining syscalls.

We now also expose relevant syscall info to userspace like the syscall
numbers and data types.

What's left to do is create an intermediary file format for defining
syscall numbers, and respect that order when defining them in C
(right now they are hardcoded to work).

Another thing that may be kind of cool is to let drivers define their
own syscalls for whatever internal use they might have. I have no clue
if this is a good idea though, we'll see.