tools: Add tools directory with a syscall generator.
We now generate syscalls based on a syscall definition file. I decided
to do this since before we had to define the SYS_* macro with the
syscall's number and then define the actual syscall in the right
position further down in the file. Now we just write the syscall
definition in an intermediary file and the syscall generator tool takes
care of everything.