vfs: Add vfs_rmfile().
We can now remove a file from a filesystem. The way this works is, when
vfs_rmfile() is called, it flags the file for deletion, and tries to
delete it.

The deletion attempt fails if there are any active references to that
vnode (the vnode has been opened as part of a file descriptor using
vfs_open). This way we don't leave any open file descriptors dangling.

Everytime vfs_close() is called on a file the reference count is
decreased, and the file is tried for deletion again, succeeding only if
there are no more references.

Only ramfs implements the underlying function necessary for deletion.