falloc: Cache falloc_one_use() state.
Use some variables to cache the last known location of free pageframes.
This massively improves the performance of the allocator, which is seen
when allocating large numbers of frames for userspace binaries.

This change is a hack and doesn't change the fact that falloc is a piece
of shit, mainly because of the fact that the pageframe pool is
statically allocated, and while this works on 32bit, on 64bit it won't.

A rework should be coming soon.