serialio: Add serialio subsystem.
The SerialIO subsystem is a modular subsystem used to register serial
devices along with their drivers. This has been constructed to allow
building of a ps2kbd.

The hitch is that the device -> driver matching is kind of awkward. It
works in two possible ways dev.major_id -> drv.major_id or dev.name ->
drv.name. I don't know if this will hold for future devices and/or buses,
or if this subsystem has been made too abstract, and will need narrowing
down.