errno: Add ERRNO_MAX define.
The motivation behind this is mmap. mmap returns a void*, and while we
could do ERR_OR_PTR(void) inside the kernel, when we return from the
syscall it gets transformed back into a void*. mmap now returns errnos
as normal values (no negatives). In userspace we can figure if mmap
errored out if the return value < ERRNO_MAX since mmap is not going to
return a map address that is < ERRNO_MAX.