fbsink+fb: Signal framebuffer takeover from userspace.
If a userspace program opens a framebuffer it will get flagged as being
taken over. In this case the kernel will not write anything to that
framebuffer.

Also add a stinky hack for writing to a framebuffer while in a user
processes' context. The framebuffer is mapped at a known location only
in the kernel. If we're coming from a process, we temporarily map the
kernel's paging struct, do our business and then go back to the original
paging struct.